+++ lib/models/entrepreneurs/ent_user.dart
@@ -1,9 +1,12 @@
 import 'package:MobileApp/models/entrepreneurs/entrepreneur_model.dart';
 
 class EntrepreneurUserModel {
-  EntrepreneurPortfolioModel entrepreneurPortfolioModel =
-      EntrepreneurPortfolioModel();
+  EntrepreneurPortfolioModel entrepreneurPortfolioModel;
   String email;
   String password;
   String username;
+
+  EntrepreneurUserModel() {
+    entrepreneurPortfolioModel = EntrepreneurPortfolioModel();
+  }
 }