+++ lib/models/entrepreneurs/entrepreneur_model.dart
@@ -6,11 +6,13 @@ class EntrepreneurPortfolioModel {
   String execSummary;
   String linkedinProfile;
   String displayImage;
+  String location;
 
   EntrepreneurPortfolioModel(
       {this.execSummary,
       this.id,
       this.uid,
+      this.location,
       this.linkedinProfile,
       this.firstName,
       this.lastName,