+++ lib/models/investors/inv_user.dart
@@ -1,8 +1,12 @@
 import 'package:MobileApp/models/investors/investor_model.dart';
 
 class InvestorUserModel {
-  InvestorPortfolioModel investorPortfolioModel = InvestorPortfolioModel();
+  InvestorPortfolioModel investorPortfolioModel;
   String username;
   String password;
   String email;
+
+  InvestorUserModel() {
+    investorPortfolioModel = InvestorPortfolioModel();
+  }
 }