+++ lib/screens/investors/saved_ven.dart
@@ -1,8 +1,7 @@
 import 'package:MobileApp/backend/entrepreneurs/portfolio_list.dart';
 import 'package:MobileApp/models/entrepreneurs/venture_model.dart';
 import 'package:MobileApp/screens/investors/inv_home_page.dart';
-import 'package:MobileApp/shared/nav_drawer.dart';
-import 'package:MobileApp/theme/app.dart';
+
 import 'package:MobileApp/theme/colors.dart';
 import 'package:flutter/material.dart';